move wirte las to m3c2 base class
